requirements.txt
asgiref==3.8.1
Django==4.1.5
python-dateutil==2.9.0.post0
python-whois==0.9.5
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
